HBX-1910: Refactor JdbcMetadataBuilder into smaller responsibilities - Move JdbcMetadataBuilder#updatePrimaryKey(...) to class 'org.hibernate.tool.internal.reveng.binder.PrimaryKeyBinder'
